removed unnecessary state
xIsNext was a state tracking whether the next move should be an X
or O.  However, this state was redundant because whether the next
move is X or O is entirely tracked by whether the currentMove
state is even or odd.  So, this state was removed and all refs to
it where changed or removed as appropriate.